# packages/contract/openapi.yaml
openapi: 3.0.3
info:
  title: amigo.sh dispatch
  version: 0.1.0
paths:
  /api/dispatch:
    post:
      summary: Create a short link (optionally with note + ttl)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  nullable: true
                note:
                  type: string
                  nullable: true
                text:
                  type: string
                  nullable: true
                ttl:
                  type: string
                  example: "7d"
                burn:
                  type: boolean
                max_views:
                  type: integer
                reply_to:
                  type: string
                  nullable: true
                emoji:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  short:
                    type: string
                  original:
                    type: string
                    nullable: true
                  note:
                    type: string
                    nullable: true
                  reply_to:
                    type: string
                    nullable: true
                  emoji:
                    type: string
                    nullable: true
            text/plain:
              schema:
                type: string
  /api/resolve/{slug}:
    get:
      summary: Resolve to JSON (url + note)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    nullable: true
                  note:
                    type: string
                    nullable: true
                  expires_at:
                    type: integer
                    nullable: true
        "404":
          description: Not found
  /api/peek/{slug}:
    get:
      summary: Peek signet + state (no burn)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Peek result
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                  gone:
                    type: boolean
                  emoji:
                    type: string
                    nullable: true
                  has_url:
                    type: boolean
  /api/burn/{slug}:
    delete:
      summary: Burn a room explicitly
      description: >
        The Nuclear Option. Destroys the room and all content immediately,
        regardless of remaining views or time.
      parameters:
        - in: path
          name: slug
          required: true
          description: The unique identifier of the room to burn
          schema:
            type: string
      responses:
        "200":
          description: Room incinerated successfully
          content:
            text/plain:
              schema:
                type: string
                example: ok
        "500":
          description: Server error during burn sequence
  /{slug}:
    get:
      summary: Resolve + redirect
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect
        "404":
          description: Not found